receipt-scanner-app/receipt-scanner-backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.12"}
psycopg = {extras = ["binary"], version = "^3.2.9"}
python-multipart = "^0.0.20"
pillow = "^11.2.1"
pytesseract = "^0.3.13"
langchain-openai = "^0.3.16"
python-dotenv = "^1.1.0"
langchain = "^0.3.25"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
